dinner_service/pyproject.toml
[project]
name = "dinner-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "email-validator>=2.3.0",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.35.0",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "pytest>=8.4.1",
]
